implemented and tested all POST, GET, DELETE, PUT endpoints for CRUD operations

+++ jobapptrack/src/main/java/com/example/jobapplicationtracker/jobapptrack/controller/JobApplicationController.java
@@ -1,6 +1,7 @@
 package com.example.jobapplicationtracker.jobapptrack.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -9,6 +10,7 @@ import com.example.jobapplicationtracker.jobapptrack.model.JobApplication;
 import com.example.jobapplicationtracker.jobapptrack.service.JobApplicationService;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/jobapplications")
@@ -27,11 +29,28 @@ public class JobApplicationController {
         return service.getAllApplications();
     }
 
+    @GetMapping("/{id}")
+    public ResponseEntity<JobApplication> getApplicationById(@PathVariable Long id){
+        Optional<JobApplication> jobApp = service.getApplicationById(id);
+        return jobApp.map(ResponseEntity::ok)
+                     .orElseGet(() -> ResponseEntity.notFound().build());
+    }
+
     @GetMapping("/status/{status}")
     public List<JobApplication> getApplicationsByStatus(@PathVariable ApplicationStatus status) {
         return service.getApplicationsByStatus(status);
     }
 
+    @GetMapping("/search")
+    public List<JobApplication> searchApplications(@RequestParam String keyword) {
+        // find keyword in company, position, notes
+        return service.getAllApplications().stream()
+                // filter and check for both caps nad lowercase
+                .filter(app -> (app.getCompany().toLowerCase().contains(keyword.toLowerCase())
+                || app.getPosition().toLowerCase().contains(keyword.toLowerCase())
+                || app.getNotes().toLowerCase().contains(keyword.toLowerCase())))
+                .toList();
+    }
 
     @PostMapping
     public JobApplication addApplication(@RequestBody JobApplication jobApplication) {
@@ -40,8 +59,14 @@ public class JobApplicationController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteApplication(@PathVariable Long id) {
-        service.deleteApplication(id);
-        return ResponseEntity.ok().build();
+        Optional<JobApplication> jobApp = service.getApplicationById(id);
+        if (jobApp.isPresent()) {
+            service.deleteApplication(id);
+            return ResponseEntity.ok().body("Job Application deleted successfully.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Job Application with id `" + id + "` not found.");
+        }
     }
 
     @PutMapping("/{id}")
